// src/components/TopBar/TopBar.Styles.d.ts
export declare const navBaseStyles: {
    header: string;
    primaryWrapper: string;
    secondaryWrapper: string;
    navInnerWrapper: string;
};
export declare const navLogoTitle: {
    logoWrapper: string;
    cuLogoLink: string;
    cuLogo: string;
    cuShield: string;
    siteTitle: string;
    oneLineHeader: string;
    twoLineHeader: string;
};
export declare const navItemStyles: {
    navWrapper: string;
    navItem: string;
    navItemChildren: string;
    navItemMoreNav: string;
    navArrow: string;
};
export declare const navDropDownStyles: {
    dropDownContainer: string;
    dropDownItems: string;
    moreMenuWrapper: string;
    moreMenuContainer: string;
    moreMenuDropDown: string;
    moreMenuParentItem: string;
    moreMenuChildItem: string;
    moreMenuParentArrow: string;
};
export declare const navAsideStyles: {
    asideWrapper: string;
    searchIcon: string;
    unorderedList: string;
    listHidden: string;
    listItemLink: string;
    megaMenuButton: string;
};
//# sourceMappingURL=TopBar.Styles.d.ts.map